+++ test/helpers/fixture.ts
@@ -9,11 +9,23 @@ export const fixtures = Object.create(null);
 
 for (const name of fixtureNames) {
     fixtures[name] = {
-        repo: (options?: GitReaderOptions) =>
-            createGitReader(
-                fileURLToPath(new URL(fixturesPath + name + '/_git', import.meta.url)),
-                options
-            ),
+        repo: (options?: GitReaderOptions) => {
+            const origReaddir = fs.promises.readdir;
+
+            try {
+                // Patch readdir method to catch issues when readdir returns unsorted entries
+                // @ts-expect-error ts(2322)
+                fs.promises.readdir = (...args: Parameters<typeof fs.promises.readdir>) =>
+                    origReaddir(...args).then((list) => list.reverse());
+
+                return createGitReader(
+                    fileURLToPath(new URL(fixturesPath + name + '/_git', import.meta.url)),
+                    options
+                );
+            } finally {
+                fs.promises.readdir = origReaddir;
+            }
+        },
         data: JSON.parse(
             fs.readFileSync(
                 fileURLToPath(new URL(fixturesPath + name + '/data.json', import.meta.url)),